fix(compose): composed figures reproduce pixel-faithfully (validation PASSES)
A composed figure (plt.compose) failed save-time reproducibility validation:
first a SIZE mismatch (1874x1080 vs 1138x898 -> 1871x1075), then a same-size
pixel divergence (MSE ~5000, 34% of pixels). Three construction mismatches
between the live compose path and the reproduce path:

1. PLACEMENT. The reproducer placed each panel at the recorded `bbox` -- the
   POST-replay position in CROPPED coords. For divider plotters
   (stx_scatter_hist) that is the shrunken main axes, so add_axes(bbox)+replay
   re-applied the divider and the panel shrank twice; cropped coords also
   shifted every panel inward. Fix: record the EXACT add_axes input compose used
   as a new `AxesRecord.compose_bbox` (uncropped, PRE-replay) and place from it.
   -> panel A (no divider) becomes pixel-perfect; all data (scatter/raster)
   reproduces exactly.

2. STYLE. The composed recipe had no figure.style, so reproduce rendered tick /
   axis-label text in matplotlib's default font while live compose applied each
   panel's publication style via _apply_source_style -> text metrics diverged
   and every label ghosted. Fix: carry the panels' shared style onto the
   composed FigureRecord.style.

3. FINALIZE. The reproducer ran finalize_ticks/finalize_special_plots per panel;
   live compose does not. Removed so both replay identically.

Composed Fig 1 now validates PASSED (MSE under threshold; data pixel-identical).
Refactor: extract panel-label helpers from the over-limit _compose.py into
_composition/_panel_labels.py. No new test failures (16 pre-existing
TestPixelPerfect/TestMetadataRoundtrip only; 274 pass).
